+++ CHANGELOG.md
@@ -6,6 +6,7 @@ New:
 
 - Language: add global modality `G`
 - Add `UPDATE!` command for in web and cli interface (by @MitchBoontjes)
+- Solving Cheryl's Sudoku using DEMO-S5
 
 Changed: